orderform: fix tifdate not showing when tif type if selected
